* 'master' of https://gitcode.com/chuguowei/ops-transformer: (856 commits)
  add FAG interception for T
  [mc2]matmul_all_reduce clean code
  修复mm2LoadDataA，mm2LoadDataB搬运为奇数时导致的越界和NAN问题
  新增AlltoAllvQuantGroupedMatmul、QuantGroupedMatmulAlltoAllv算子
  codecheck clean
  添加pa nz无rope场景下的d的拦截
  fix GMMSwigluQunatV2 A8W8 multiTensor
  新增支持MhcSinkhorn算子
  Dispatch No Quant Cast To HIFP8
  fix bsa document aclnn sample code
  [PLZ]aiv直驱dpu代码
  A5 QSFAP
  BlockSparseAttention document & README修改
  rm qkv warning
  support cv11 for gmm
  新增 mhcPost 算子
  950上moeinitrouting v3 算子hif8量化模板 降低代码重复率
  修复allToAllMatmul场景，smoothQuant功能
  alltoallmatmul&matmulalltoall算子获取整型属性风险整改
  dav3510检查aic和aic核数比
  ...

+++ moe/moe_token_unpermute_with_ep/op_host/moe_token_unpermute_with_ep_tiling.cpp
@@ -103,8 +103,8 @@ static inline ge::graphStatus MoeTokenUnpermuteWithEpInputParamCheck(
     auto dataTensor0 = context->GetInputTensor(0);
     auto dataTensor1 = context->GetInputTensor(1);
     auto nodeName = context->GetNodeName();
-    const int* tmpTopK = context->GetAttrs()->GetAttrPointer<int>(0);
-    int inputTopK = static_cast<int64_t>(*tmpTopK);
+    const int64_t* tmpTopK = context->GetAttrs()->GetAttrPointer<int64_t>(0);
+    int inputTopK = *tmpTopK;
 
     OP_CHECK_IF(
         inputTopK < 1, OP_LOGE(nodeName, "input num TopK cannot less than 1."),
@@ -171,8 +171,8 @@ static inline void Init(
     param.input.isUnpermute = isUnpermute;
 
     auto attrPtr = context->GetAttrs();
-    const int* tmpTopK = attrPtr->GetAttrPointer<int>(UNPERMUTE_WITH_EP_ARRT_TOPK);
-    int inputTopK = static_cast<int64_t>(*tmpTopK);
+    const int64_t* tmpTopK = attrPtr->GetAttrPointer<int64_t>(UNPERMUTE_WITH_EP_ARRT_TOPK);
+    int inputTopK = *tmpTopK;
     auto rangePtr = attrPtr->GetAttrPointer<gert::ContinuousVector>(UNPERMUTE_WITH_EP_ARRT_RANGE);
     if (rangePtr != nullptr) {
         OP_CHECK_IF(
